requirements.txt
gradio==4.42.0
loguru==0.6.0
numpy==2.1.1
Pillow==10.4.0
qdrant_client==1.11.1
torch==2.4.0
tqdm==4.66.2
